ToDo_List/todo_project/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
typing-extensions==4.12.2
